Database.Drasil

Description

Re-export database types and functions to simplify external use.

Documentation

module Drasil.Database.Chunk

module Drasil.Database.UID

module Drasil.Database.TypedUIDRef

Core database types and functions.

data ChunkDB

Drasil's knowledge database.

empty :: ChunkDB

An empty ChunkDB.

fromList :: IsChunk a => [a] -> ChunkDB

Create a ChunkDB from a list of chunks. This will insert all chunks into the database from the list, from left to right.

registered :: ChunkDB -> [UID]

Query the ChunkDB for all registered chunks (by their UIDs).

typesRegistered :: ChunkDB -> [TypeRep]

Enumerate all types registered in the ChunkDB.

size :: ChunkDB -> Int

Get the number of chunks registered in the ChunkDB.

isRegistered :: UID -> ChunkDB -> Bool

Check if a UID is registered in the ChunkDB.

findUnused :: ChunkDB -> [UID]

Filter the ChunkDB for chunks that are not needed by any other chunks. These are the only chunks that can safely be removed from the database, though we do not include this functionality.

find :: Typeable a => UID -> ChunkDB -> Maybe a

Find a chunk by its UID in the ChunkDB.

findOrErr :: forall a. Typeable a => UID -> ChunkDB -> a

Find a chunk by its UID in the ChunkDB, throwing a hard error if it is not found.

findAll :: forall a. IsChunk a => ChunkDB -> [a]

Find all chunks of a specific type in the ChunkDB.

findAll' :: TypeRep -> ChunkDB -> [UID]

Find all chunks of a specific type in the ChunkDB, returning their UIDs rather than the chunks themselves.

dependants :: UID -> ChunkDB -> Maybe [UID]

Find all chunks that depend on a specific one.

dependantsOrErr :: UID -> ChunkDB -> [UID]

Find all chunks that depend on a specific one, throwing a hard error if the dependency chunk is not found.

findTypeOf :: UID -> ChunkDB -> Maybe TypeRep

Find the type of a chunk by its UID.

insert :: IsChunk a => a -> ChunkDB -> ChunkDB

Insert a chunk into the ChunkDB if it is sensibly to do so (i.e., does not depend on itself and is not a ChunkDB). We temporarily allow chunks to overwrite other ones, but we warn when this happens.

insertAll :: IsChunk a => [a] -> ChunkDB -> ChunkDB

Insert a list of chunks into a ChunkDB.

Temporary functions

insertAllOutOfOrder12 :: (IsChunk a, IsChunk b, IsChunk c, IsChunk d, IsChunk e, IsChunk f, IsChunk g, IsChunk h, IsChunk i, IsChunk j, IsChunk k) => ChunkDB -> [a] -> [b] -> [c] -> [d] -> [e] -> [f] -> [g] -> [h] -> [i] -> [j] -> [k] -> [Reference] -> ChunkDB

Insert 12 lists of unique chunk types into a ChunkDB, assuming the input ChunkDB does not already contain any of the chunks from the chunk lists.

NOTE: Ignores management of dependancies related to LabelledContents and References.

refTable :: ChunkDB -> Map UID Reference

refFind :: UID -> ChunkDB -> Reference

Find a Reference by its UID, throwing an error if it is not found.

refbyTable :: ChunkDB -> Map UID [UID]

refbyLookup :: UID -> Map UID [UID] -> [UID]

Find what chunks reference a given UID.

traceTable :: ChunkDB -> Map UID [UID]

traceLookup :: UID -> Map UID [UID] -> [UID]

Query a chunk for to what chunks it refers to.

type ChunkType = String

type DumpedChunkDB = Map ChunkType [UID]

A dumped representation of a ChunkDB, mapping chunk types (Strings) to lists of UIDs.

dumpChunkDB :: ChunkDB -> DumpedChunkDB

data TypedUIDRef typ

UID references that contain information about the type of data the UID refers to, useful for type-safe dereferencing.

mkRef :: IsChunk t => t -> TypedUIDRef t

Create a TypedUIDRef to a chunk.

typedFind :: IsChunk t => TypedUIDRef t -> ChunkDB -> Maybe t

Find a chunk by its typed UID reference.

typedFindOrErr :: IsChunk t => TypedUIDRef t -> ChunkDB -> t

Find a chunk by its typed UID reference, erroring if not found.