Database.Drasil.SystemInformation

Documentation

data SystemInformation where

Data structure for holding all of the requisite information about a system to be used in artifact generation.

Constructors

SI

Fields

:: (CommonIdea a, Idea a, Idea b, HasName c, Quantity e, Eq e, MayHaveUnit e, Quantity f, MayHaveUnit f, Concept f, Eq f, Quantity h, MayHaveUnit h, Quantity i, MayHaveUnit i, HasUID j, Constrained j)

=> { _sys :: a

, _kind :: b

, _authors :: [c]

, _purpose :: d

, _quants :: [e]

, _concepts :: [f]

, _instModels :: [InstanceModel]

, _datadefs :: [DataDefinition]

, _configFiles :: [String]

, _folderPath :: FilePath

, _inputs :: [h]

, _outputs :: [i]

, _defSequence :: [Block QDefinition]

, _constraints :: [j]

, _constants :: [QDefinition]

, _sysinfodb :: ChunkDB

, _usedinfodb :: ChunkDB

, refdb :: ReferenceDB

} -> SystemInformation

data Block a

for listing QDefinitions in SystemInformation.

Constructors

Coupled a a [a]

Parallel a [a]

citeDB :: SystemInformation -> BibRef

Helper for extracting a bibliography from the system information.

data ReferenceDB

Database for internal references. Contains citations and referrable concepts.

citationsFromBibMap :: BibMap -> [Citation]

Create sorted citations from a bibliography.

citationDB :: Lens' ReferenceDB BibMap

rdb :: BibRef -> [ConceptInstance] -> ReferenceDB

Smart constructor for creating a reference database from a bibliography and concept instances.

type RefMap a = Map UID (a, Int)

Database for maintaining references. The Int is that reference's number. Maintains access to both num and chunk for easy reference swapping between number and shortname/refname when necessary (or use of number if no shortname exists).

simpleMap :: HasUID a => [a] -> RefMap a

Constructor that makes a RefMap from things that have a UID.

conceptDB :: Lens' ReferenceDB ConceptMap

instModels :: Lens' SystemInformation [InstanceModel]

datadefs :: Lens' SystemInformation [DataDefinition]

configFiles :: Lens' SystemInformation [String]

folderPath :: Lens' SystemInformation FilePath

inputs :: HasInputs c => Lens' c [(QuantityDict, Maybe (RealInterval Expr Expr))]

Provides a Lens that holds a QuantityDict and maybe constraints.

defSequence :: Lens' SystemInformation [Block QDefinition]

constraints :: Constrained c => Lens' c [ConstraintE]

Provides a Lens to the Constraints.

constants :: Lens' SystemInformation [QDefinition]

sysinfodb :: Lens' SystemInformation ChunkDB

usedinfodb :: Lens' SystemInformation ChunkDB