Drasil.Database

Description

Re-export database types and functions to simplify external use.

Documentation

data Chunk

A piece of reusable knowledge, with an internal identifier (UID), possibly dependant on other chunks.

type IsChunk a = (HasUID a, HasChunkRefs a, Typeable a)

Constraint for anything that may be considered a valid chunk type.

class HasChunkRefs a where

All chunks should expose what chunks they reference/rely on, so that we can test ChunkDBs to ensure all presupposed chunks are already registered.

Methods

chunkRefs :: a -> Set UID

mkChunk :: IsChunk a => a -> Chunk

Create a Chunk, ensuring that Chunks are never placed within Chunks.

unChunk :: Typeable a => Chunk -> Maybe a

Open chunks, casting them to a specific type.

chunkType :: Chunk -> TypeRep

Ask a Chunk for the type of data it codifies.

Core database types and functions.

data ChunkDB

Drasil's knowledge database.

empty :: ChunkDB

An empty ChunkDB.

fromList :: IsChunk a => [a] -> ChunkDB

Create a ChunkDB from a list of chunks. This will insert all chunks into the database from the list, from left to right.

registered :: ChunkDB -> [UID]

Query the ChunkDB for all registered chunks (by their UIDs).

typesRegistered :: ChunkDB -> [TypeRep]

Enumerate all types registered in the ChunkDB.

size :: ChunkDB -> Int

Get the number of chunks registered in the ChunkDB.

isRegistered :: UID -> ChunkDB -> Bool

Check if a UID is registered in the ChunkDB.

findUnused :: ChunkDB -> [UID]

Filter the ChunkDB for chunks that are not needed by any other chunks. These are the only chunks that can safely be removed from the database, though we do not include this functionality.

find :: Typeable a => UID -> ChunkDB -> Maybe a

Find a chunk by its UID in the ChunkDB.

findOrErr :: forall a. Typeable a => UID -> ChunkDB -> a

Find a chunk by its UID in the ChunkDB, throwing a hard error if it is not found.

findAll :: forall a. IsChunk a => ChunkDB -> [a]

Find all chunks of a specific type in the ChunkDB.

findAll' :: TypeRep -> ChunkDB -> [UID]

Find all chunks of a specific type in the ChunkDB, returning their UIDs rather than the chunks themselves.

dependants :: UID -> ChunkDB -> Maybe [UID]

Find all chunks that depend on a specific one.

dependantsOrErr :: UID -> ChunkDB -> [UID]

Find all chunks that depend on a specific one, throwing a hard error if the dependency chunk is not found.

findTypeOf :: UID -> ChunkDB -> Maybe TypeRep

Find the type of a chunk by its UID.

insert :: IsChunk a => a -> ChunkDB -> ChunkDB

Insert a chunk into the ChunkDB if it is sensible to do so (i.e., does not depend on itself, is not a ChunkDB, and does not overwrite another chunk).

insertAll :: IsChunk a => [a] -> ChunkDB -> ChunkDB

Insert a list of chunks into a ChunkDB.

Temporary functions

insertAllOutOfOrder11 :: (IsChunk a, IsChunk b, IsChunk c, IsChunk d, IsChunk e, IsChunk f, IsChunk g, IsChunk h, IsChunk i, IsChunk j, IsChunk k) => ChunkDB -> [a] -> [b] -> [c] -> [d] -> [e] -> [f] -> [g] -> [h] -> [i] -> [j] -> [k] -> ChunkDB

Insert 11 lists of unique chunk types into a ChunkDB, assuming the input ChunkDB does not already contain any of the chunks from the chunk lists.

type ChunkType = String

type DumpedChunkDB = Map ChunkType [UID]

A dumped representation of a ChunkDB, mapping chunk types (Strings) to lists of UIDs.

dumpChunkDB :: ChunkDB -> DumpedChunkDB

data UID

A UID is a 'unique identifier' for things that we will put into our database of information. We use a newtype wrapper to make sure we are only using UIDs where desired.

class HasUID c where

The most basic item: having a unique identifier key, here a UID.

Methods

uid :: Getter c UID

Provides a unique id for internal Drasil use.

mkUid :: String -> UID

Smart constructor for UIDs from raw Values.

nsUid :: String -> UID -> UID

Nest a UID under a namespace.

(+++) :: HasUID a => a -> String -> UID

Append a suffix to a thing with a UID and get the resulting UID.

(+++.) :: UID -> String -> UID

Append a suffix to a UID.

(+++!) :: (HasUID a, HasUID b) => a -> b -> UID

Merge the UIDs of two chunks through simple concatenation.

showUID :: HasUID a => a -> String

Get something's UID as a Value.

data TypedUIDRef typ

UID references that contain information about the type of data the UID refers to, useful for type-safe dereferencing.

mkRef :: IsChunk t => t -> TypedUIDRef t

Create a TypedUIDRef to a chunk.

typedFind :: IsChunk t => TypedUIDRef t -> ChunkDB -> Maybe t

Find a chunk by its typed UID reference.

typedFindOrErr :: IsChunk t => TypedUIDRef t -> ChunkDB -> t

Find a chunk by its typed UID reference, erroring if not found.