Drasil.Database.Chunk

Documentation

data Chunk

A piece of reusable knowledge, with an internal identifier (UID), possibly dependant on other chunks.

type IsChunk a = (HasUID a, HasChunkRefs a, Typeable a)

Constraint for anything that may be considered a valid chunk type.

class HasChunkRefs a where

The majority of chunks will relate other chunks in some way. In other words, the majority of our chunks *depend* on others. HasChunkRefs is meant as a way to capture what things a chunk *directly* refers to (i.e., depends on directly).

Methods

chunkRefs :: a -> Set UID

mkChunk :: IsChunk a => a -> Chunk

Create a Chunk, ensuring that Chunks are never placed within Chunks.

unChunk :: Typeable a => Chunk -> Maybe a

Open chunks, casting them to a specific type.

chunkType :: Chunk -> TypeRep

Ask a Chunk for the type of data it codifies.