Drasil.Database.UIDRef

UID References

data UIDRef typ

A reference to another chunk through its UID, with a type variable to retain information about the original chunk's type, useful for type-safe dereferencing.

hide :: IsChunk t => t -> UIDRef t

Create a UIDRef to a chunk.

unhide :: IsChunk t => UIDRef t -> ChunkDB -> Maybe t

Find a chunk by a UIDRef.

unhideOrErr :: IsChunk t => UIDRef t -> ChunkDB -> t

Find a chunk by a UIDRef, erroring if not found.

data UnitypedUIDRef

A variant of UIDRef without type information about the chunk being referred to, effectively treating chunks as being "unityped."

hideUni :: IsChunk t => t -> UnitypedUIDRef

Create a UnitypedUIDRef to a chunk.

unhideUni :: IsChunk t => UnitypedUIDRef -> ChunkDB -> Maybe t

Find a chunk by its UnitypedUIDRef.

unhideUniOrErr :: IsChunk t => UnitypedUIDRef -> ChunkDB -> t

Find a chunk by its UnitypedUIDRef, erroring if not found.