Database.Drasil

Documentation

data ChunkDB

Our chunk databases. Must contain all maps needed in an example. In turn, these maps must contain every chunk definition or concept used in its respective example, else an error is thrown.

Constructors

CDB SymbolMap TermMap ConceptMap UnitMap TraceMap RefbyMap DatadefnMap InsModelMap GendefMap TheoryModelMap ConceptInstanceMap SectionMap LabelledContentMap ReferenceMap

type RefbyMap = Map UID [UID]

A reference map, used to hold a UID and where it is referenced (UIDs).

type TraceMap = Map UID [UID]

A traceability map, used to hold the relation between one UID and a list of other UIDs.

type UMap a = Map UID (a, Int)

The misnomers below (for the following Map types) are not actually a bad thing. We want to ensure data can't be added to a map if it's not coming from a chunk, and there's no point confusing what the map is for. One is for symbols + their units, and the others are for what they state.

asOrderedList :: UMap a -> [a]

Gets an ordered list of a from any a that is of type UMap.

cdb :: (Quantity q, MayHaveUnit q, Idea t, Concept c, IsUnit u) => [q] -> [t] -> [c] -> [u] -> [DataDefinition] -> [InstanceModel] -> [GenDefn] -> [TheoryModel] -> [ConceptInstance] -> [Section] -> [LabelledContent] -> [Reference] -> ChunkDB

Smart constructor for chunk databases. Takes in the following:

[Quantity] (for SymbolMap),

NamedIdeas (for TermMap),

Concepts (for ConceptMap),

Units (something that IsUnit for UnitMap),

DataDefinitions (for DatadefnMap),

InstanceModels (for InsModelMap),

GenDefns (for GendefMap),

TheoryModels (for TheoryModelMap),

ConceptInstances (for ConceptInstanceMap),

Sections (for SectionMap),

LabelledContents (for LabelledContentMap).

collectUnits :: Quantity c => ChunkDB -> [c] -> [UnitDefn]

Gets the units of a Quantity as UnitDefns.

conceptMap :: Concept c => [c] -> ConceptMap

Smart constructor for a ConceptMap.

conceptinsLookup :: UID -> ConceptInstanceMap -> ConceptInstance

Looks up a UID in the concept instance table. If nothing is found, an error is thrown.

conceptinsTable :: Lens' ChunkDB ConceptInstanceMap

dataDefnTable :: Lens' ChunkDB DatadefnMap

datadefnLookup :: UID -> DatadefnMap -> DataDefinition

Looks up a UID in the datadefinition table. If nothing is found, an error is thrown.

defResolve :: ChunkDB -> UID -> ConceptChunk

Looks up a UID in the definition table from the ChunkDB. If nothing is found, an error is thrown.

gendefLookup :: UID -> GendefMap -> GenDefn

Looks up a UID in the general definition table. If nothing is found, an error is thrown.

gendefTable :: Lens' ChunkDB GendefMap

generateRefbyMap :: TraceMap -> RefbyMap

Translates a traceability map into a reference map.

refResolve :: UID -> ReferenceMap -> Reference

Looks up a UID in the reference table from the ChunkDB. If nothing is found, an error is thrown.

refTable :: Lens' ChunkDB ReferenceMap

idMap :: HasUID a => [a] -> Map UID (a, Int)

General smart constructor for making a UMap out of anything that has a UID.

termMap :: Idea c => [c] -> TermMap

Smart constructor for a TermMap.

insmodelLookup :: UID -> InsModelMap -> InstanceModel

Looks up a UID in the instance model table. If nothing is found, an error is thrown.

insmodelTable :: Lens' ChunkDB InsModelMap

labelledconLookup :: UID -> LabelledContentMap -> LabelledContent

Looks up a UID in the labelled content table. If nothing is found, an error is thrown.

labelledcontentTable :: Lens' ChunkDB LabelledContentMap

refbyLookup :: UID -> RefbyMap -> [UID]

Trace a UID to referenced UIDs.

refbyTable :: Lens' ChunkDB RefbyMap

sectionLookup :: UID -> SectionMap -> Section

Looks up a UID in the section table. If nothing is found, an error is thrown.

sectionTable :: Lens' ChunkDB SectionMap

symbResolve :: ChunkDB -> UID -> QuantityDict

Looks up a UID in the symbol table from the ChunkDB. If nothing is found, an error is thrown.

termResolve :: ChunkDB -> UID -> IdeaDict

Looks up a UID in the term table from the ChunkDB. If nothing is found, an error is thrown.

termTable :: ChunkDB -> TermMap

theoryModelLookup :: UID -> TheoryModelMap -> TheoryModel

Looks up a UID in the theory model table. If nothing is found, an error is thrown.

theoryModelTable :: Lens' ChunkDB TheoryModelMap

traceLookup :: UID -> TraceMap -> [UID]

Trace a UID to related UIDs.

traceMap :: [(UID, [UID])] -> TraceMap

Smart constructor for a TraceMap given a traceability matrix.

traceTable :: Lens' ChunkDB TraceMap

ccss :: [Sentence] -> [DisplayExpr] -> ChunkDB -> [DefinedQuantityDict]

Gets a list of defined quantities (DefinedQuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

ccss' :: [Sentence] -> [DisplayExpr] -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

combine :: Sentence -> ChunkDB -> [DefinedQuantityDict]

Combines the functions of vars and concpt to create a list of DefinedQuantityDicts from a Sentence.

getIdeaDict :: Sentence -> ChunkDB -> [IdeaDict]

Gets a list of ideas (IdeaDict) from a Sentence in order to print.

vars :: DisplayExpr -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDict) from an equation in order to print.

data Block a

for listing QDefinitions in SystemInformation.

Constructors

Parallel a [a]

type RefMap a = Map UID (a, Int)

Database for maintaining references. The Int is that reference's number. Maintains access to both num and chunk for easy reference swapping between number and shortname/refname when necessary (or use of number if no shortname exists).

data ReferenceDB

Database for internal references. Contains citations and referrable concepts.

data SystemInformation where

Data structure for holding all of the requisite information about a system to be used in artifact generation.

Constructors

SI

Fields

:: (CommonIdea a, Idea a, Idea b, HasName c, Quantity e, Eq e, MayHaveUnit e, Quantity f, MayHaveUnit f, Concept f, Eq f, Quantity h, MayHaveUnit h, Quantity i, MayHaveUnit i, HasUID j, Constrained j)

=> { _sys :: a

, _kind :: b

, _authors :: [c]

, _purpose :: d

, _quants :: [e]

, _concepts :: [f]

, _instModels :: [InstanceModel]

, _datadefs :: [DataDefinition]

, _configFiles :: [String]

, _folderPath :: FilePath

, _inputs :: [h]

, _outputs :: [i]

, _defSequence :: [Block QDefinition]

, _constraints :: [j]

, _constants :: [QDefinition]

, _sysinfodb :: ChunkDB

, _usedinfodb :: ChunkDB

, refdb :: ReferenceDB

} -> SystemInformation

citeDB :: SystemInformation -> BibRef

Helper for extracting a bibliography from the system information.

rdb :: BibRef -> [ConceptInstance] -> ReferenceDB

Smart constructor for creating a reference database from a bibliography and concept instances.

simpleMap :: HasUID a => [a] -> RefMap a

Constructor that makes a RefMap from things that have a UID.

citationDB :: Lens' ReferenceDB BibMap

conceptDB :: Lens' ReferenceDB ConceptMap

sysinfodb :: Lens' SystemInformation ChunkDB

folderPath :: Lens' SystemInformation FilePath