Database.Drasil.ChunkDB

Description

Defines types and functions to create a chunk database within Drasil.

Types

ChunkDB

Main database type

data ChunkDB

Our chunk databases. Must contain all maps needed in an example. In turn, these maps must contain every chunk definition or concept used in its respective example, else an error is thrown.

Constructors

CDB

Fields

symbolTable :: SymbolMap

termTable :: TermMap

conceptChunkTable :: ConceptMap

_unitTable :: UnitMap

_dataDefnTable :: DatadefnMap

_insmodelTable :: InsModelMap

_gendefTable :: GendefMap

_theoryModelTable :: TheoryModelMap

_conceptinsTable :: ConceptInstanceMap

_citationTable :: CitationMap

_labelledcontentTable :: LabelledContentMap

_refTable :: ReferenceMap

_traceTable :: TraceMap

_refbyTable :: RefbyMap

data TermAbbr

Constructors

TermAbbr

Fields

longForm :: NP

shortForm :: Maybe String

data DomDefn

Constructors

DomDefn

Fields

domain :: [UID]

definition :: Sentence

Maps

Exported for external use.

type RefbyMap = Map UID [UID]

A reference map, used to hold a UID and where it is referenced (UIDs).

type TraceMap = Map UID [UID]

A traceability map, used to hold the relation between one UID and a list of other UIDs.

type UMap a = Map UID (a, Int)

The misnomers below (for the following Map types) are not actually a bad thing. We want to ensure data can't be added to a map if it's not coming from a chunk, and there's no point confusing what the map is for. One is for symbols + their units, and the others are for what they state.

Functions

Constructors

cdb' :: (Quantity q, MayHaveUnit q, Concept q, Concept c, IsUnit u) => [q] -> [IdeaDict] -> [c] -> [u] -> [DataDefinition] -> [InstanceModel] -> [GenDefn] -> [TheoryModel] -> [ConceptInstance] -> [LabelledContent] -> [Reference] -> [Citation] -> ChunkDB

Smart constructor for chunk databases. Takes in the following:

[Quantity] (for SymbolMap),

NamedIdeas (for TermMap),

Concepts (for ConceptMap),

Units (something that IsUnit for UnitMap),

DataDefinitions (for DatadefnMap),

InstanceModels (for InsModelMap),

GenDefns (for GendefMap),

TheoryModels (for TheoryModelMap),

ConceptInstances (for ConceptInstanceMap),

LabelledContents (for LabelledContentMap). Creates an empty ChunkDB without basic data. cdb in MetaDatabase.Drasil should be preferred over this function as it includes the basic data needed for Drasil.

idMap :: HasUID a => String -> [a] -> Map UID (a, Int)

General smart constructor for making a UMap out of anything that has a UID.

symbolMap :: (Quantity c, MayHaveUnit c, Concept c) => [c] -> SymbolMap

Smart constructor for a SymbolMap.

termMap :: Idea c => [c] -> TermMap

Smart constructor for a TermMap.

conceptMap :: Concept c => [c] -> ConceptMap

Smart constructor for a ConceptMap.

unitMap :: IsUnit u => [u] -> UnitMap

Smart constructor for a UnitMap.

traceMap :: [(UID, [UID])] -> TraceMap

Smart constructor for a TraceMap given a traceability matrix.

generateRefbyMap :: TraceMap -> RefbyMap

Translates a traceability map into a reference map.

Lookup Functions

asOrderedList :: UMap a -> [a]

Gets an ordered list of a from any a that is of type UMap.

collectUnits :: Quantity c => ChunkDB -> [c] -> [UnitDefn]

Gets the units of a Quantity as UnitDefns.

termResolve :: (NP -> Maybe String -> c) -> ChunkDB -> UID -> c

Search for _any_ chunk that is an instance of Idea and process its "term" and abbreviation.

termResolve' :: ChunkDB -> UID -> TermAbbr

Find a chunks "term" and abbreviation, if it exists.

defResolve :: ([UID] -> Sentence -> c) -> ChunkDB -> UID -> c

Looks up a UID in all tables with concepts from the ChunkDB. If nothing is found, an error is thrown.

symbResolve :: ChunkDB -> UID -> DefinedQuantityDict

Looks up a UID in the symbol table from the ChunkDB. If nothing is found, an error is thrown.

traceLookup :: UID -> TraceMap -> [UID]

Trace a UID to related UIDs.

refbyLookup :: UID -> RefbyMap -> [UID]

Trace a UID to referenced UIDs.

datadefnLookup :: UID -> DatadefnMap -> DataDefinition

Looks up a UID in the datadefinition table. If nothing is found, an error is thrown.

insmodelLookup :: UID -> InsModelMap -> InstanceModel

Looks up a UID in the instance model table. If nothing is found, an error is thrown.

gendefLookup :: UID -> GendefMap -> GenDefn

Looks up a UID in the general definition table. If nothing is found, an error is thrown.

theoryModelLookup :: UID -> TheoryModelMap -> TheoryModel

Looks up a UID in the theory model table. If nothing is found, an error is thrown.

conceptinsLookup :: UID -> ConceptInstanceMap -> ConceptInstance

Looks up a UID in the concept instance table. If nothing is found, an error is thrown.

refResolve :: UID -> ReferenceMap -> Reference

Looks up a UID in the reference table from the ChunkDB. If nothing is found, an error is thrown.

Lenses

unitTable :: Lens' ChunkDB UnitMap

traceTable :: Lens' ChunkDB TraceMap

refbyTable :: Lens' ChunkDB RefbyMap

citationTable :: Lens' ChunkDB CitationMap

dataDefnTable :: Lens' ChunkDB DatadefnMap

insmodelTable :: Lens' ChunkDB InsModelMap

gendefTable :: Lens' ChunkDB GendefMap

theoryModelTable :: Lens' ChunkDB TheoryModelMap

conceptinsTable :: Lens' ChunkDB ConceptInstanceMap

labelledcontentTable :: Lens' ChunkDB LabelledContentMap

refTable :: Lens' ChunkDB ReferenceMap

Helpers

addCdb :: ChunkDB -> ChunkDB -> ChunkDB

defResolve' :: ChunkDB -> UID -> DomDefn